[Codegen][AMDGPU] Drop backend reverts, emergency RDNA4 lowering fix (#21906)
The lowering for amdgpu.lds_barrier upstream is incorrect in the face of
the backend changes whose reverts this commit drops. The lowering will
be fixed in the near future. Meanwhile, stick with gpu.barrier
(incurring the full __synchthreads() semantics and thus potentially
introduceing unwanted fencing on global memory) for RDNA4 until the
issue upstream is resolved.

This may incur performance penalties in some cases.