[CPU][DT] Add codegen support for broadcast/dequant -> matmul dispatch. (#21911)
The encoding materialization is moved before RematerializeParallelOps
pass, because it can make matmul ops (in generic op form) have more than
two input operands and the materialization pattern does not recognize
the case. Instead of having the checks in the pass, the revision runs
the materialization pass earlier, which is an easier and reasonable fix.

It also adds LVMCPUTileToVectorSize pass the mmt4d pipeline, so the
producer can be tiled to target vector size, which avoids large vectors
and stack allocations.

Fixes https://github.com/iree-org/iree/issues/21866

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>