Don't inline immutable globals with non-util dialect attrs. (#21986)
`mlir::Dialect::materializeConstant` does not have a way for us to pass
these and preserve them on the new op and we don't (yet) have an attr
interface that lets us control the behavior. For now any non-util
dialect attr will prevent inlining. Inlining is most useful early on for
CSE and local folding but given that the predominate dialect attr near
the frontend is an explicitly-specified `stream.affinity` and that's
really bad to drop the conservative behavior is fine for now.

(we do have `IREE::Util::InliningPolicyAttrInterface` we could add some
methods to, but the way FoldGlobalsPass works today doesn't really allow
it to have the control it needs - whenever FoldGlobalsPass does finally
get rewritten properly it should use it)

Fixes #21946.