[CPU] Switch matmul config to use linalg::LinalgOp interface. (#21954)
IREE tends to generalize linalg operations, which improves the fusion.
In this context, a matmul op can be in a linalg.generic op form. The
revision updates the interface op to `linalg::LinalgOp` which provides
more interface methods. The check is done by
`linalg::isaContractionOpInterface`.

It does not break any existing lit tests; it adapts one named op test to
linalg.generic form.

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>