[CPU] Populate to_elements unrolling patterns in LLVM conversion. (#22010)
It is a follow-up to https://github.com/iree-org/iree/pull/21851. The
CPU backend also needs the patterns, which is not surprising. It was not
detected in the first place because a test did not exist. It only
happens in dynamic gather+attention cases.

The test is failing on VMVX and SPIR-V backend. Thus, the test is only
added to CPU and hip for now.

Fixes https://github.com/iree-org/iree/issues/22007

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>