[DispatchCreation] Add option to infer split-reduction sizes for batchnorm (#21731)
This adds a pass to infer split-reduction sizes for outer-reduction
operations (disabled by default, enabled by
`--iree-dispatch-creation-enable-split-reduction`).

This is targeted at the batch normalization operation, which performs
reductions over the outer dimensions on a tensor. Some examples:
```mlir
%9 = linalg.generic {
  indexing_maps = [
    affine_map<(d0, d1, d2, d3) -> (d0, d2, d3, d1)>,
    affine_map<(d0, d1, d2, d3) -> (d1)>
  ],
  iterator_types = ["reduction", "parallel", "reduction", "reduction"]
} ins(%6 : tensor<128x24x48x384xf32>) outs(%8 : tensor<384xf32>) {
^bb0(%in: f32, %out: f32):
  %22 = arith.addf %in, %out : f32
  linalg.yield %22 : f32
} -> tensor<384xf32>
```
```mlir
%11 = linalg.generic {
  indexing_maps = [
    affine_map<(d0, d1, d2, d3) -> (d0, d2, d3, d1)>,
    affine_map<(d0, d1, d2, d3) -> (d1)>,
    affine_map<(d0, d1, d2, d3) -> (d1)>
  ],
  iterator_types = ["reduction", "parallel", "reduction", "reduction"]
} ins(%6, %10 : tensor<128x24x48x384xf32>, tensor<384xf32>) outs(%8 : tensor<384xf32>) {
^bb0(%in: f32, %in_5: f32, %out: f32):
  %22 = arith.subf %in, %in_5 : f32
  %23 = arith.mulf %22, %22 : f32
  %24 = arith.addf %23, %out : f32
  linalg.yield %24 : f32
} -> tensor<384xf32>
```

The pass targets a reduction tile size of `1024`, with the tile size
rounded up until it evenly divides the iteration domain. For example, a
reduction of `tensor<2050xf32> -> tensor<1xf32>` would be tiled with a
split-reduction tile size of `[1025]`, resulting in two parallel
iterations of `tensor<1025xf32> -> tensor<1xf32>` reductions, and a
`tensor<2xf32> -> tensor<1xf32>` merge step.

When multiple reduction dimensions are present, the inner reduction
dimensions are tiled first, to produce contiguous slices for each tile.
For example a `tensor<8x260xf32> -> tensor<1xf32>` reduction is tiled to
2 parallel `tensor<4x260xf32> -> tensor<1xf32>` reductions.