[LinalgExt] Support map_scatter decomposition with strided memrefs (#21952)
Support decomposing map_scatter ops into vector.scatter when the output
buffer of the map_scatter is not collapsible. Sometimes the source
memref layout is strided (typically coming from dynamic sizes + offset
in a subspan), and we need to handle the case. The decomposition will
now create a `memref.reinterpret_cast` on the output buffer to convert
it to the flat shape, and the strides are folded into the index
transformation region of the map_scatter before decomposing.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>