[Encoding] Propagate layout encodings through tensor.cast ops (#21970)
Implements encoding propagation for LayoutAttr encodings upwards through
`tensor.cast` operations. We will start seeing models with these layout
encodings in the input coming from the front end, so we need to handle
this propagation in order to open up fusions for the set_encoding ops.

The PR also refactors the propagation op interface implementations to
make it simpler to add the new interface implementation for tensor.cast
and future ops that can be considered "encoding castable", meaning their
propagation implementation just involves casting the op types to the
encoded type.

The `propagateEncoding` interface method is also updated to take an
`OpOperand *` instead of `OpResult`, because it carries the operand
number, making it less ambiguous. The documentation for the interface
method is updated with a more detailed description of what the
`propagateEncoding` function is meant to do.

The implementation for `LayoutAttr::isSerialized` is updated to actually
check that all of its layouts are themselves serialized, instead of
unconditionally returning true. The encoding resolvers for CPU/GPU/VMVX
also need an implementation for `isSerialized`, which is added in this
PR too. The default implementation for the `isSerialized` interface
method now returns `false` instead of throwing an assertion error so
that the compiler does not crash when calling the method on an attribute
that does not implement it.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>