Disable data-tiling flag by default and refresh the CPU docs. (#21935)
We made the decision that turns data-tiling on by default for CPU only
in the past; we realized that it is a mistake today.

The motivation was reducing questions like how do we get good
performance in CPU backend; people can discover the data-tiling
technique easily. However, it's a very involved to enable data-tiling by
default, especially when some backends do not support/use data-tiling.
That means data-tiling is not ready for O2 atm. Even for CPU backend, it
can result in additional memory uses that we don't have good control
yet, which is another data point that data-tiling is not ready for O2.

To address the discoverability issue, the revision updates the CPU docs.
It is the start point for new users, and they can copy the recommended
commands from the doc. For the active existing users, they're tagged in
the PR and they will discover the change in the future release note. For
the hidden existing users, they can only find it in the release note.

The revision updates all the external_tests suite to use the
`--iree-opt-data-tiling` flag.

Preview of CPU doc:
https://hanhanw.github.io/iree/guides/deployment-configurations/cpu/#compile-a-program

Preview of data-tiling optimization option:
https://hanhanw.github.io/iree/reference/optimization-options/#data-tiling-iree-opt-data-tiling-off

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>