[GPU][DT] Add data-tiling resolver by default.  (#22074)
The resolver was not added for a few main reasons:
* It took some time to finish the work. It was not easy to make it
default when data-tiling was on by default.
* There are three approaches (none, pad, data-tiling). Since data-tiling
is off by default, `none` is redundant in the context. We prefer
`data-tiling` over `pad`, so it makes sense to make `data-tiling` be
used by default.

It is a step toward reducing required flags for data-tiling. The issue
was identified in https://github.com/iree-org/iree/issues/22073. It is
what we want, but it is not the proper fix to the issue.

It does not matter if we've already materialized the encoding with a
resolver. E.g.,

```mlir
#encoding = #iree_encoding.layout<[#iree_gpu.gpu_encoding_resolver<configuration = {encoding_info = {innerDimsPos = [0, 1], innerTileSizes = [64, 16], outerDimsPerm = [0, 1], swizzle = {expandShape = [[["CrossThread", 16 : i16], ["CrossIntrinsic", 4 : i16]], [["Internal", 16 : i16]]], permutation = [1, 0, 2]}}}>]>
#encoding1 = #iree_encoding.layout<[#iree_gpu.gpu_encoding_resolver<configuration = {encoding_info = {innerDimsPos = [0, 1], innerTileSizes = [64, 16], outerDimsPerm = [0, 1], swizzle = {expandShape = [[["CrossThread", 4 : i16], ["CrossThread", 16 : i16]], [["Internal", 16 : i16]]], permutation = [0, 1, 2]}}}>]>
```

However, it does not apply to all the cases. We still need the resolver
to materialize unserialized encodings. It is a missing feature in
current implementation. The correct implementation is checking if an
encoding is serialized or not:
* If so, do the materialization through interface methods.
* If not, grab the resolver from executable target, if any, and
materialize the encodings.

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>