[GPU] Fix bug in shared memory computation for scaled intrinsics (#22095)
The shared memory calculation was not taking into account both of the K
dimensions (K and the block dim) when computing shared memory tiles, so
it was underestimating shared memory use for schedules, and selecting
tile sizes that do not fit in shared memory. This fixes the issue by
properly setting the `kSize` field in the `GPUMMASchedule` to the
product of all intrinsic K dimensions.

The lit tests for TileAndFuse configs are updated, because the shapes in
the tests were too small to detect this bug. The shapes are made large
to test the logic.

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>