Fix the Windows build: portably set environment variable PYTHONPATH. (#22136)
This fixes the breakage in
https://github.com/iree-org/iree/actions/runs/18060514263/job/51396076814#step:9:11510
which was caused by setting the PYTHONPATH environment variable in
Bash-like syntax on Windows. Using `cmake -E env` should be the portable
way to do that.

I tried the alternative of setting the working directory for this
comment instead of PYTHONPATH, but no luck with that.

ci-extra: windows_x64_msvc

Signed-off-by: Benoit Jacob <jacob.benoit.1@gmail.com>