[Codegen][AMDGPU] Fix incorrect canonical map for MXFP RHS scales (#22162)
The "canonical" shape of the inner tiles for a group-scaled inner tiled
matrix multiply is
(M, Ko, Kb) * (M, Ko) x (Ko, Kb, N) * (Ko, N) => (M, N) where the
`permutations` attribute of the inner-tiled operation gives permutations
from that canonical shape to the shape of the actual inner tiles.

The method for converting a scalid mfma represented as a linalg.generic
to an inner-tiled operation would use the affine maps on the
linalg.geniric to compute these permutation attributes. However, for the
RHS scales in particular, the utility function was given a (N, Ko)
canonical shape and not a (Ko, N) one, resulting in the generation of
incorrect indexing code.

This commit fixis the issue.

Fixes #22104