E2E MXFP4 matmul tests (#22170)
This further refactors the e2e matmul test python generator to split the
code generation into separate Python files, so that the verbose MX code
generation is its own separate file. This surfaced a preexisting problem
in dependency tracking: the test rules needed to depend not only on the
main .py file but also on others. Fixed by depending on all .py files in
the current source dir.

The approach in these tests is inspired by
https://github.com/qedawkins/fp4_matmul_testing. That is: the tests
generate MLIR code scaling and converting the inputs from MXFP4 to FP32
so that the existing FP32 test harness can be used without changes.

Two tests are added: without and with data-tiling. The latter is not
actually taking effect at the moment but will when
https://github.com/Max191/scaled-matmul-encoding-materialization-full-layout
is merged, and the tests are already known to pass when rebased on that.

Fixes #22067

---------

Signed-off-by: Benoit Jacob <jacob.benoit.1@gmail.com>