[LinalgExt] Introduce linalg_ext.exp_reduction (#21761)
In the online softmax algorithm, there are two reductions: the
calculation of the maximum term, and the accumulation of the sum term
which is sum of e^(value[i] - max). Combined, we call this reduction
"exponential reduction".

The fusion of the two reductions is non-trivial, as it influences the
error accumulation in the floating point numbers. By identifying
exponential reduction, we enable the fusion and subsequently easier
optimization of kernels that use it.

Support for exp_reduction will be added incrementally across multiple
PRs:
- Add the iree_linalg_ext.exp_reduction op definition and verification
logic, along with roundtrip and invalid IR tests (this PR)
- Add TilingInterface implementation.
- Add DecomposeExpReductionPass, which will decompose the op into
elemwise linear algebra ops, along with relevant tests.
- Add e2e tests.

Co-authored-by: Kunwar Grover <groverkss@gmail.com>

---------

Signed-off-by: Ivan Ho <ivan.ho@cl.cam.ac.uk>