[DataTiling] Introduce DataTiledMMAInterfaceAttr (#22098)
Introduces the `DataTiledMMAInterfaceAttr`, and implements it for
`DataTiledMMAAttr`. This PR is mostly about moving the existing
implementation for `DataTiledMMAAttr`, so the functionality is mostly
the same and no new tests are needed.

The new interface is added in order to simplify the data tiling
materialization and codegen implementations for data tiled scaled
matmuls. The idea behind the interface is that the
`InnerTileDescAttrInterface` implementation, and the data tiling
materialization implementation can both be derived from the pack +
TileSwizzle layouts of the attribute's root operation.

To do this, the PR extends the `TileSwizzle::Dim` definition to add an
explicit distribution size. This is necessary, because there are some
special cases where the tensor shape of TileSwizzle does not match the
distribution shape of the TileSwizzle. This is the case for the WMMA
intrinsics, or when we have subgroup unrolling along both M and N,
because we need to distribute a larger number of threads than the
cumulative size of the TileSwizzle. With this additional field, the
TileSwizzle can now fully describe both the layout and distribution of
the layout, and the implementation for the `InnerTileDescAttrInterface`
can be derived from just the TileSwizzle, and some very simple
additional interface functions.

A shared implementation of the majority of the
`InnerTileDescAttrInterface` is included as part of a new helper class
in IREEGPUAttrs.td called `IREEGPU_DataTiledMMAInnerTileDescAttr`, which
defines the interface implementations using the shared implementation in
IREEGPUInterfaces.cpp. Any new attribute inheriting from
`DataTiledMMAInterfaceAttr` should probably inherit from this class in
order to simplify the interface implementations.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>