[Codegen] Fix premature return in iree_codegen.inner_tiled verifier (#22183)
There was a bug in the iree_codegen.inner_tiled verifier that caused the
function to return success before making all the necessary checks. This
PR removes the early return, and adds some `invalid` lit tests to
properly test the verifier. The PR also removes 2 redundant checks that
are already present in the verifier for the DestinationStyleOpInterface.

This also updates an invalid pack_to_intrinsics test, which is now
caught by the verifier.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>