Adding a LiftCFGToSCFPass. (#22101)
This is similar to the upstream ControlFlowToSCFPass and reuses the
upstream infrastructure but focuses it on IREE::Util ops. This allows us
to support unreachable ops today (they don't exist upstream) and as we
add new SCF-like ops (such as ones that support tied operands) we'll be
able to lift into those.

Fairly exhaustive tests are included (some from upstream but adjusted to
our ops) but the pass is not currently used. There's a few more PRs
before we get to activating it in the compilation pipeline as we have
quite a few places that have only been nominally tested with scf (they
fail or may work but by disabling critical optimizations). We also
insert CFG control flow in a few places and need to change those emit
SCF (ideally we are CFG-free in the core flow/stream pipelines).

(there's one known issue here that is fixed in #22102 - without
`util.scf.unreachable` we can't fully lift infinite loops)