Promote iree-opt-data-tiling to pipeline options. (#22295)
There are two paths for data-tiling. One starts from global optimization
phase and the other starts from dispatch creation phase. The former one
is the suggested path, since the latter one is just getting used
recently.

The revision adds a new `iree-global-opt-data-tiling` flag that allows
users to explicitly start the data-tiling from global optimization
phase; it promotes the existing flag to an umbrella flag for
data-tiling, so users can start data-tiling with the suggested path. It
is intended to be NFC for most users in terms of command line flags. The
`iree-opt-data-tiling` used to be identical to the new
`iree-global-opt-data-tiling`, and it selects the global optimization
path when it is set.

Note that the pipeline flag has higher priority. I.e., it overrides the
other two flags if it is set as true.

Doc preview:
https://hanhanw.github.io/iree/reference/optimization-options/#data-tiling-iree-opt-data-tiling-off

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>